--- engine/src/scriptwrapper.h
@@ -1,11 +1,0 @@
-
-#ifndef SCRIPTWRAPPER_H
-#define SCRIPTWRAPPER_H
-
-#ifdef QT_QML_LIB
-  #include "scriptv4.h"
-#else
-  #include "script.h"
-#endif
-
-#endif
